feat(org-jira): add default start time and worklog duration
Added two customize variables: `org-jira-default-log-start-time` and
`org-jira-default-worklog-duration`, to set the starting time and the
duration of logging work in Jira.

Refactored the calculation of the worklog start time in the
`org-jira-mini-read-worklog-data` function to make efficient use of the
new customizable start time and default duration.

Modified functions `org-jira-last-weekday-at` and
`org-jira-mini-read-woklog-data` to calculate a default start time based
on user preferences and handle the edge-case scenario when the current
day is a weekend or the current time is earlier than the desired start
time.

In such cases, the function will automatically find the closest previous
weekday to assign as the default start time.

Finally, changed the `Hours:` prompt to use the user's specified
`org-jira-default-worklog-duration`.